requirements.txt
django
djangorestframework
django-cors-headers
requests
celery
redis
django-celery
djangorestframework-jwt
django[argon2]
psycopg2